Removed feature flags as development is complete
