+++ src/components/ListingCard/listingCard.module.css
@@ -1,55 +1,20 @@
-.listing {
-  align-items: center;
-  background-color: var(--whitebase---ffffff);
-  border-radius: 16px;
-  box-shadow: 0px 4px 4px 
-#00000040;
-  display: flex;
-  flex-direction: column;
-  height: 274px;
-  overflow: hidden;
-  position: relative;
-  width: 225px;
-  margin: 30px 30px;
-  cursor: pointer;
-}
-
-.details {
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  height: 160px;
-  margin-left: 2px;
-  padding: 11px 0;
-  width: 225px;
-}
-
-.title {
-  letter-spacing: 0;
-  color: var(--black);
-  text-align: center;
-  line-height: 40px;
-  margin-left: 1px;
-  min-height: 40px;
-  min-width: 124px;
-  white-space: nowrap;
-}
-
-.description {
-  color: var(--black);
-  font-weight: 400;
-  line-height: 23px;
-  margin-top: 11px;
-  min-height: 69px;
-  text-align: center;
-  width: 171px;
+.card {
+    box-shadow: 0px 4px 4px #00000040;
+    width: 205px;
 }
 
 .avatar {
-  background-position: 50% 50%;
-  background-size: cover;
-  border-radius: 200px;
-  height: 90px;
-  margin-top: 24px;
-  width: 89px;
+    height: 100px;
+    width: 100px;
+    border-radius: 50%;
+    object-fit: cover;
+    box-shadow: 0px 4px 4px #00000040;
+    margin-bottom: 4px;
+}
+
+.listingImg {
+    height: 100px;
+    width: 100px;
+    object-fit: cover;
+    margin-bottom: 4px;
 }
