--- src/components/ListingCard/newListingCard.module.css
@@ -1,20 +1,0 @@
-.card {
-    box-shadow: 0px 4px 4px #00000040;
-    width: 205px;
-}
-
-.avatar {
-    height: 100px;
-    width: 100px;
-    border-radius: 50%;
-    object-fit: cover;
-    box-shadow: 0px 4px 4px #00000040;
-    margin-bottom: 4px;
-}
-
-.listingImg {
-    height: 100px;
-    width: 100px;
-    object-fit: cover;
-    margin-bottom: 4px;
-}
